Extract grouped query evaluation
